Use libc to get the current time
This improves performance by ~26% based on the criterion "generation"
benchmark and 10% based on the "string_generation" benchmark. The
smaller improvement for the string generation makes sense as most of
that time is dominated by creating the string.

ID generation still benchmarks much slower, 1.5x, than the BSON gem's
implementation. I think a lot of that will be coming from the Ruby
interop now.