Freeze the ObjectId class
Freezing the class allows for faster generation of IDs since the class
is thread-safe and can be shared between threads. The ObjectId struct is
also now frozen since it's not mutated after creation. I didn't notice
any performance improvement from this but it seems like a good idea to
do.